requirements.txt
tweepy
picamera
pillow
